Change services, rename helpers -> validators, formatted all files
