+++ src/categories/services/changeCategoryName.ts
@@ -1,6 +1,6 @@
 import { MySubclassedDexie } from "../../pages";
 import { Category } from "../types/category";
-import { categoryNameExists } from "../helpers/categoryExists";
+import { categoryNameExists } from "../validators/categoryExists";
 
 export const changeCategoryName = async (
   db: MySubclassedDexie,
@@ -10,5 +10,6 @@ export const changeCategoryName = async (
   if (categoryNameExists(existingCategories, name)) {
     throw new Error("This category exists!");
   }
+  
   return db.categories.update(uuid, { name: name.trim() });
 };
