+++ src/categories/validators/categoryExists.ts
